refactor: Replace hash-based navigation with proper SvelteKit routing
Replace manual hash change handling with proper SvelteKit route parameters:
- Create /countries/[section] route structure
- Add server-side validation with redirect for invalid sections
- Maintain backward compatibility with legacy hash URLs
- Update dropdown to use programmatic navigation
- Replace manual event listeners with SvelteKit's goto()

Benefits:
- Proper browser history support
- Server-side route validation
- Cleaner, more maintainable code
- Better SEO with real URLs
- Follows SvelteKit best practices
